move observeResize to shared
