+++ client/web/src/util/dom.ts
@@ -1,32 +1,10 @@
-import { head, isArray } from 'lodash'
+import { isArray } from 'lodash'
 import { useMemo } from 'react'
 import { Observable } from 'rxjs'
 import { catchError, debounceTime, map } from 'rxjs/operators'
 
+import { observeResize } from '@sourcegraph/shared/src/util/dom'
 import { useObservable } from '@sourcegraph/wildcard'
-
-/**
- * An Observable wrapper around ResizeObserver
- */
-export const observeResize = (target: HTMLElement): Observable<ResizeObserverEntry | undefined> => {
-    let animationFrameID: number
-
-    return new Observable(function subscribe(observer) {
-        const resizeObserver = new ResizeObserver(entries => {
-            // Move `ResizeObserver` measurements into a RAF to avoid the "ResizeObserver loop limit exceeded" error.
-            // See the thread for background info: https://github.com/WICG/resize-observer/issues/38
-            animationFrameID = window.requestAnimationFrame(() => {
-                observer.next(head(entries))
-            })
-        })
-        resizeObserver.observe(target)
-
-        return function unsubscribe() {
-            window.cancelAnimationFrame(animationFrameID)
-            resizeObserver.disconnect()
-        }
-    })
-}
 
 interface ObserveQuerySelectorInit {
     /**
