+++ client/web/src/views/components/view/content/chart-view-content/charts/line/components/scroll-box/ScrollBox.tsx
@@ -2,8 +2,7 @@ import classNames from 'classnames'
 import React, { useEffect, useState } from 'react'
 
 import { isFirefox } from '@sourcegraph/shared/src/util/browserDetection'
-
-import { observeResize } from '../../../../../../../../../util/dom'
+import { observeResize } from '@sourcegraph/shared/src/util/dom'
 
 import styles from './ScrollBox.module.scss'
